refine storage proof benchmarks and skip v8 changes
